pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["KateRomanova <kate.romanova@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
black = "^24.8.0"
isort = "^5.13.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.26.0"
coverage = "^7.6.1"
drf-yasg = "^1.21.7"
stripe = "^10.8.0"
forex-python = "^1.8"
celery = "^5.4.0"
redis = "^5.0.8"
django-celery-beat = "^2.7.0"
setuptools = "^74.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
